+++ packages/bridge-controller/src/types.ts
@@ -95,7 +95,7 @@ export type TokenAmountValues = {
   /**
    * The amount of the token
    *
-   * @example "1000000000000000000"
+   * @example "1.005"
    */
   amount: string;
   /**